restructured coreconcepts component
